Quick Subnet Restriction
Quick and dirty subnet restriction. I run two containers, one that is accessible externally and one that is "internal" only. The external container uses authentication and encryption but the internal one is essentially open to all. The problem is that I do not want to rely entirely on cloud provider's ACLs. Since docker seems to bypass firewalld and set it's own iptable entries to allow access to these ports, it's helpful to restrict access to subnets not routable across the internet.