feat: add support for setting the timezone with $TZ
This allows users of the container to simply set the timezone using `TZ` environment variable.

On host systems like talos linux the timezone of the kubernetes worker is fixed by definition to UTC, so mounting in the host's `/etc/localtime` would make no difference. Setting the environment variable causes glibc to automatically lookup the timezone information, but without the tzdata package installed  it cannot obtain the information needed to understand the timezone requested, and fails back to displaying the time in UTC.